requirements.txt
pytorch=1.7.1
cudatoolkit
transformers=3.5.1
gitpython
